requirements.txt
Django==4.1.7
mysqlclient==2.1.1
sqlparse==0.4.3
Pillow==9.5.0
pandas==2.0.1
django-widget-tweaks==1.4.12
django-phonenumber-field==7.1.0
phonenumbers==8.13.11
celery==5.2.7
redis==4.6.0
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.2.1
pytest-factoryboy==2.5.1
